task_tracker/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
markdown==3.0.1
django-filter==2.0.0